requirements.txt
# Backend dependencies for FastAPI
fastapi[standard]
uvicorn>=0.15.0
qdrant-client>=1.1.1
python-dotenv>=0.19.0
redis>=4.0.1
pydantic>=2.5.1
sqlalchemy>=1.4.48
psycopg2-binary>=2.9.5
asyncpg>=0.27.0
websockets

# Authentication and environment loading dependencies
bcrypt==4.0.1
python-jose==3.3.0

boto3

faker>=20.0.0

# Google Gemini
google-genai
assemblyai
numpy>=1.23.0
google-generativeai>=0.3.1
gTTS>=2.5.4

pymongo
celery[redis]
flower

python-jose[cryptography]
passlib[bcrypt]
python-multipart
alembic
email-validator
jinja2
aiofiles
httpx
itsdangerous
python-dateutil
pytz
python-slugify

pinecone
pymilvus[model]>=2.4.0

# Langchain Eco System
langchain
langchain_community
langchain-google-genai
langgraph
httpx>=0.27.0

speechrecognition>=3.10.0
gTTS>=2.5.4
markdown>=3.4.0
beautifulsoup4>=4.12.0
elevenlabs>=1.0.0
pydantic[email]>=2.0.0
requests>=2.31.0
soundfile>=0.12.1
pillow>=10.0.0
azure-storage-blob
azure-identity

# P2 Backend Dependencies
openai>=1.0.0
faiss-cpu>=1.7.4
pydantic-settings>=2.0.0